requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
finnhub-python==2.4.27
idna==3.11
numpy==2.4.2
pandas==3.0.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
urllib3==2.6.3

